+++ src/__experimental__/components/decimal/decimal.spec.js
@@ -7,7 +7,7 @@ import { testStyledSystemMargin } from "../../../__spec_helper__/test-utils";
 import Decimal from "./decimal.component";
 import Textbox from "../../../components/textbox/textbox.component";
 import Label from "../../../__internal__/label";
-import FormFieldStyle from "../form-field/form-field.style";
+import FormFieldStyle from "../../../__internal__/form-field/form-field.style";
 
 // These have been written in a way that we can change our testing library or component implementation with relative
 // ease without having to touch the tests.
