+++ src/components/button-toggle-group/button-toggle-group.component.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styledSystemPropTypes from "@styled-system/prop-types";
 
-import FormField from "../../__experimental__/components/form-field";
+import FormField from "../../__internal__/form-field";
 import ButtonToggleGroupStyle from "./button-toggle-group.style";
 import RadioButtonMapper from "../../__experimental__/components/radio-button/radio-button-mapper.component";
 import ValidationIcon from "../../__internal__/validations/validation-icon.component";
