+++ src/components/configurable-items/configurable-items.style.js
@@ -5,7 +5,7 @@ import {
   ConfigurableItemRowIconStyle,
   ConfigurableItemRowStyle,
 } from "./configurable-item-row/configurable-item-row.style";
-import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
+import StyledFormField from "../../__internal__/form-field/form-field.style";
 import StyledCheckbox from "../checkbox/checkbox.style";
 
 import baseTheme from "../../style/themes/base";
