+++ src/components/form/form.style.js
@@ -2,7 +2,7 @@ import styled, { css, keyframes } from "styled-components";
 import PropTypes from "prop-types";
 
 import { space } from "styled-system";
-import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
+import StyledFormField from "../../__internal__/form-field/form-field.style";
 import { StyledFieldset } from "../../__internal__/fieldset/fieldset.style";
 
 import StyledButton from "../button/button.style";
