+++ src/components/grouped-character/grouped-character.spec.js
@@ -3,7 +3,7 @@ import { mount } from "enzyme";
 
 import { testStyledSystemMargin } from "../../__spec_helper__/test-utils";
 import GroupedCharacter from "./grouped-character.component";
-import FormFieldStyle from "../../__experimental__/components/form-field/form-field.style";
+import FormFieldStyle from "../../__internal__/form-field/form-field.style";
 import Label from "../../__internal__/label";
 
 const mountComponent = (props) => mount(<GroupedCharacter {...props} />);
