+++ src/components/numeral-date/numeral-date.spec.js
@@ -10,8 +10,8 @@ import {
   testStyledSystemMargin,
 } from "../../__spec_helper__/test-utils";
 import StyledInputPresentation from "../../__internal__/input/input-presentation.style";
-import FormFieldStyle from "../../__experimental__/components/form-field/form-field.style";
-import FormField from "../../__experimental__/components/form-field";
+import FormFieldStyle from "../../__internal__/form-field/form-field.style";
+import FormField from "../../__internal__/form-field";
 import { rootTagTest } from "../../utils/helpers/tags/tags-specs";
 import Label from "../../__internal__/label";
