+++ src/components/numeral-date/numeral-date.style.js
@@ -2,8 +2,8 @@ import styled, { css } from "styled-components";
 import StyledValidationIcon from "../../__internal__/validations/validation-icon.style";
 import StyledIconSpan from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import { baseTheme } from "../../style/themes";
-import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
 import StyledInputPresentantion from "../../__internal__/input/input-presentation.style";
+import StyledFormField from "../../__internal__/form-field/form-field.style";
 import StyledIcon from "../icon/icon.style";
 
 export const StyledNumeralDate = styled.div`
