+++ src/components/textbox/textbox.d.ts
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { SpaceProps } from "styled-system";
 import { ValidationPropTypes } from "../../__internal__/validations";
-import { FormFieldSize } from "../../__experimental__/components/form-field/form-field";
+import { FormFieldSize } from "../../__internal__/form-field/form-field";
 
 export interface CommonTextboxProps extends ValidationPropTypes {
   /** Breakpoint for adaptive label (inline labels change to top aligned). Enables the adaptive behaviour when set */
